+++ idea/src/org/jetbrains/kotlin/idea/refactoring/introduce/extractClass/ui/KotlinExtractInterfaceDialog.kt
@@ -54,6 +54,15 @@ class KotlinExtractInterfaceDialog(
                 extractableMemberInfos,
                 getInterfaceContainmentVerifier { selectedMembers }
         ) {
+            override fun isMemberEnabled(memberInfo: KotlinMemberInfo): Boolean {
+                if (!super.isMemberEnabled(memberInfo)) return false
+
+                val member = memberInfo.member
+                return !(member.hasModifier(KtTokens.INLINE_KEYWORD) ||
+                         member.hasModifier(KtTokens.EXTERNAL_KEYWORD) ||
+                         member.hasModifier(KtTokens.LATEINIT_KEYWORD))
+            }
+
             override fun isAbstractEnabled(memberInfo: KotlinMemberInfo): Boolean {
                 if (!super.isAbstractEnabled(memberInfo)) return false
                 val member = memberInfo.member